requirements.txt
Pillow==9.1.1
django==4.0.5
django-ckeditor==6.4.2
PyMySQL==1.0.2